add metrics about the I/O Queue
There are way too many I/O Queue metrics, so let's move them to their own
dashboard. Usually one will be looking at just a couple of metrics at once, but
that's because only some classes usually play an important role in the load.

Since the reason we have too many metrics is that we have a lot of classes,
there really isn't a way to simplify this for the main dashboard unless we leave
out really important information.

For each of the classes, 3 out of the 4 metrics are exposed:
 - delay, meaning for how long requests sit in the queue
 - bandwidth, self-explanatory
 - IOPS, self-explanatory. In conjunction with bandwidth can be
   used to find out the average size of requests.

Signed-off-by: Glauber Costa <glommer@scylladb.com>