grafana: Use resolution of 1/1
When I set scraping interval to 1s I see graphs "jiggle" in a way that
historical data points change values. After looking closer at this it
turns out that graphs alternate on refreshes between showing samples
from only odd and only even timestamp values. This is because of
intervalFactor being set to 1/2, even when auto-tuned step is small.

Setting "resolution" parameter for queries from 1/2 to 1/1.

The problem may still happen when the step is larger than scrape
period, which can happen when viewing a long time range. There the
solution would be to set the step to 1s. But I am not sure we want to
do this by default, as it would make grafana unusuable when viewing
long time ranges due to amount of data which would need to be fetched.

Refs #45