add options for prometheus and grafana ports
We sometimes have the need to monitor many deployments at once.
Instead of having to kill the current containers, it would be useful
to let them live alongside each other in different ports.

To do that, we need grafana and prometheus to be able to listen to
different ports and have different names.

Signed-off-by: Glauber Costa <glommer@scylladb.com>