simplify prometheus probing
We are currently using the private IP for the prometheus docker container,
in proxy mode. But because we always export the prometheus port - by
design - it's always http accessible from grafana and we can use direct
mode in localhost IP.

By doing this we don't have to do any discovery. This is a more robust
method: sometimes when the box reboots or any other change is made, the
restart order will change the internal IP addresses of the containers
and monitoring will stop working until we manually fix them.

With direct access, for as long as the exposed port is always the same,
it will always work.

Signed-off-by: Glauber Costa <glommer@scylladb.com>