bring back proxy mode
Local mode, albeit much simpler, doesn't work in remote installations.
In those, the browser will probe for prometheus directly and at
localhost it will be nowhere to be found.

We can still have multi-port mode working in proxy mode by making sure
we probe the correct container and noticing that we always have to
connect to its port 9090, and not the locally bound port.

Fixes #115

Signed-off-by: Glauber Costa <glommer@scylladb.com>