move referred files to a more specific path
Right now we are copying our files into /etc/prometheus. However, for
existing prometheus installations, that directory may already exist, and
we will have no control over its contents.

In particular, "node_exporter_servers.yml" is a quite obvious name for
someone using node_exporter, and copying our files there can override
existing ones.

This is important because although the user can put prometheus.yml
anywhere - and point to it with -config.file <something>, the files that
are referred to in the main configuration file are expected to be listed
in absolute paths.

In our container, that we control, we can leave the main prometheus.yaml
in /etc/prometheus, to avoid having to pass extra options, but we will
move the remaining files to /etc/scylla.d/prometheus. This way both the
docker environment and externally-installed environments can use the
same files, pointing to the same path, without fear of overriding
anything.

Signed-off-by: Glauber Costa <glauber@scylladb.com>