change "per Cluster" dashboard into an overview dashboard
The per Cluster separation doesn't make sense anymore since we have
added templating to the other dashboards - now we can get cluster
aggregates in every dashboard.

The value this dashboard still brings is that it only includes some
metrics, as opposed to a sea of metrics and it's great for an initial
overview.

We already have plans of starting to separate metrics by subsystem,
with an overview dashboard providing the entry point. Let this be
the first step.

Signed-off-by: Glauber Costa <glauber@scylladb.com>