enhance per server dashboard with useful metrics
First, to open up some real estate in the dashboard, I am removing the total
operations graph that sits on top: it is now redundant, because the user
can aggregate per Cluster in the entire dashboard.

With this out of the way, I am doing the following:
- Adding a graph with foreground CPU utilization. That is the CPU used by
  request processing, excluding compaction, flushes and other things. The reason
  for that is that users are usually scared of spikes. Even if we tell them that
  spikes are fine because they are the result of isolatable background processes,
  it is hard to *prove* that without further analysis. This graph will help.

- time spent in violations: A lot of the latency issues we have, specially in
  higher percentiles come from task quota violations. We have a metric for this
  now and it will help us correlate latency spikes in time

- Client connections: in the past few months, this is *THE* top metric we
  have been looking at to detect problems. It harms us a lot that it is not
  part of the main dashboard.

In the process of doing the above, I am also doing my best to document the new
graphs. The text will appear in the tooltip in the top left corner of the graph.

Caveat: The request per sec graph, where it is, becomes too narrow. I don't
know how to make it wider.

Signed-off-by: Glauber Costa <glauber@scylladb.com>