start-all.sh: check that scylla/node target files exists before starting
Currently `start-all.sh` will silently ignore inexistent or non-file
paths passed to `-s` and `-n`. This leads to an empty dashboard that has
`N/A` available nodes and the user trying to find the error in their
setup when in fact they just had a typo in the `start-all.sh` command
line.
When the paths passed to `-s` or `-n` don't point to a file, refuse to
start and print an error.