Add a variable for the datasource
There are situations in complex deployments where we can be monitoring different
clusters that are in different prometheus servers.

What we have been usually doing in situations like that is having many different
stacks, each with its own grafana/prometheus pair. But it would be useful to be
able to have the same entry point for it.

With this patch, the datasource becomes configurable as a grafana variable. We
can then choose which datasource we'll monitor, here called an "datasource"
and seamlessly flip between them.

Signed-off-by: Glauber Costa <glauber@scylladb.com>