Fix chart legend formats
Having the same static string for many graph series is not
very helpful as you can't tell the source instance when you
spot an anomaly in the graph.

Let's format the legend with instance and device for metrics
that are not dependent on `by` template parameter.

For metrics that change the grouping based on the `by` template
parameter we just use empty legend format since the tags that
generate series in the graph vary. With empty legend format,
Grafana includes all the information needed using its default
format.