create external directories if they don't exist
I verified today that if the prometheus data directory does not exist,
docker itself will try to create it.  However the directory will then be
created with the wrong permissions and as a result prometheus fails to
start if the code is being executed as a user.

Create the directory ourselves to prevent this issue. I am also doing
the same for grafana since it is bound to have the same issue.

Signed-off-by: Glauber Costa <glauber@scylladb.com>