podmand: workaround for podman pod support
rootless containers in podman can comunicate in two ways, the prefered
(but currently have bug and not working) is with a shared pod.

The alternative is using a host ip.
Until the bug in podman is solved and in a release, this patch uses an
arbitrary ip address of the host (the host can have multiple interfaces
anyone would do).

(cherry picked from commit 45ebfdc4782ba360fb60edbe3844ca9801f15a57)