Support the --stack command line option
It is sometimes usefull to run a secondary stack on the same machine.
This involved having different set of ports.
To make this task easier, there is now an additional command line option
for start-all.sh and kill-all.sh, the --stack id.

id could be 1, 2, 3, or 4.

This will run a minimize secondary stack on a secondary set of port
numbers.

Fixes #2164

Signed-off-by: Amnon Heiman <amnon@scylladb.com>