start-all.sh passing directory to archive
This patch makes the use of archive and the master branch easier.

Both usages are for advanced users that works on historical data.

After this change the archive command line flag will get a directory as
a parameter.

It will look for the scylla-monitoring scylla.txt in that directory, if
found it will use it to start the monitoring stack for the given
version.

Additional changes, parameter validation, if a long command line flag is
passed without a value, it will generate an error.

Running with the master branch will generate the dashboards for you.
Note that it requires python.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>