scylla-cql: treat Logged BATCH as CQL warning
Using a multi partition Logged BATCH is considered bad practice.  It
does not provide atomicity guarantees, and reduces performance, as a
request sent to a random node, not the right replica or shard.

This patch moves the logged batch panel to the cql optimization part and
adds a gauge next to it.

Fixes #2567
Signed-off-by: Amnon Heiman <amnon@scylladb.com>