scylla-ks: Add a keyspace table
For clusters with multiple keyspaces, it's sometimes hard to find the
most relevent one.

This patch adds a table with all available keyspace it include a quick
navigation and indication to how many tables and active tables are in
each keyspace.

This is mostly useful for users who keeps old unused keyspaces.

Fixes #2569

Signed-off-by: Amnon Heiman <amnon@scylladb.com>