types.json: Combine Nodes stats to a single panel
This patch combines the 3 nodes related stats panel, to a single one.
If everything is ok, it would just show the number of nodes in the
cluster.
If there are joining or leaving nodes it would also show that number.
And if there are unreachable nodes, it will show that.

Fixes #2609

Signed-off-by: Amnon Heiman <amnon@scylladb.com>