prometheus.consul.yml: limit extra labels to one metrics
It's common practice to use extra labels to pass node and cluster
specific information.

As those label are static per instance it's a waste of resources.
This patch uses relabel_config to drop a list of labels, all but the one
on scylla_scylladb_current_version.

Limiting to the extra labels to one per instance

Signed-off-by: Amnon Heiman <amnon@scylladb.com>