Add compress success
This patch adds the compression ratio success.
In the overview dashboard, timeouts are now part of the cluster status.
Disk usage replaces the number of tables and when the metric is
available, an indication to the compression success is shown in
percentage.

The compression success was also added to the keyspace dashboard for in
the tables table.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>